+++ BALSAMIC/utils/constants.py
@@ -637,11 +637,9 @@ BIOINFO_TOOL_ENV = {
     "strelka": "varcall_py27",
     "manta": "varcall_py27",
     "cnvkit": "varcall_cnvkit",
-    "delly": "varcall_delly"
-}
-
-BIOINFO_TOOL_SUBMODULE = {
-    "ascat": "ascatngs"
+    "delly": "varcall_delly",
+    "ascatNgs": "ascatngs",
+    "sentieon": "sentieon"
 }
 
 REPORT_MODEL = {